requirements_prot1.txt
absl-py==0.7.0
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
imageio==2.4.1
imageio-ffmpeg==0.2.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
moviepy==0.2.3.5
numpy==1.16.1
Pillow==5.4.1
protobuf==3.6.1
requests==2.21.0
scipy==1.2.0
six==1.12.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
tqdm==4.30.0
urllib3==1.24.1
Werkzeug==0.14.1